requirements.txt
botbuilder-integration-aiohttp>=4.14.0
SQLAlchemy==1.4.44
pytz==2021.1
python-dotenv==0.21.0
PyMySQL==1.0.2
apscheduler==3.9.1
pymongo==4.2.0
pymongo[srv]
requests
google-cloud-secret-manager
pytz